requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
spotipy==2.23.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
SpeechRecognition==3.10.0
pydub==0.25.1
colorama==0.4.6
requests==2.31.0
